Fix parser factory method return types (#62)
